+++ packages/ai/src/global.ts
@@ -1,4 +1,4 @@
-import { ProviderV2 } from '@ai-sdk/provider';
+import { ProviderV3 } from '@ai-sdk/provider';
 import { LogWarningsFunction } from './logger/log-warnings';
 
 // add AI SDK default provider to the globalThis object
@@ -11,7 +11,7 @@ declare global {
    *
    * @see https://ai-sdk.dev/docs/ai-sdk-core/provider-management#global-provider-configuration
    */
-  var AI_SDK_DEFAULT_PROVIDER: ProviderV2 | undefined;
+  var AI_SDK_DEFAULT_PROVIDER: ProviderV3 | undefined;
 
   /**
    * The warning logger to use for the AI SDK.